Cleanup and split to platform specific files
